+++ src/widget/text_widget.rs
@@ -18,7 +18,6 @@
 use graphics;
 use graphics::character::CharacterCache;
 use graphics::draw_state::DrawState;
-use graphics::types::Color;
 use graphics::*;
 use opengl_graphics::{GlGraphics, GlyphCache, TextureSettings};