+++ RELEASE_NOTES.md
@@ -8,6 +8,8 @@
 - Layout Manager redesign is working properly.
 - Created layout manager that is part of the widget store. (#150)
 - HorizontalLayoutManager code completed. (#58)
+- Redoing simple demo to handle better layout.
+- Fixed bug in BoxWidget where border exceeded drawing bounds, and did not fill properly after resize.
 
 ## 0.3.4